refactor: 색상 변수 구조를 @layer base (Tailwind v3) → @theme (Tailwind v4) 형식으로 변경 (#62)
