+++ src/components/ui/ReviewRating.tsx
@@ -29,7 +29,7 @@ export function ReviewRating({
   };
 
   return (
-    <div className='flex items-center gap-2'>
+    <div className='flex gap-2 items-center'>
       <Rating
         readonly={readOnly}
         initialRating={rating}
@@ -38,7 +38,7 @@ export function ReviewRating({
         emptySymbol={<FaStar size={size} color={inactiveColor} />}
         fullSymbol={<FaStar size={size} color={activeColor} />}
       />
-      <span className='text-sm text-(--color-primary-500-60)'>{rating.toFixed(1)} / 5</span>
+      <span className='text-sm text-primary-500-60'>{rating.toFixed(1)} / 5</span>
     </div>
   );
 }
